requirements.txt
fastapi
uvicorn[standard]
redis>=5.0
python-dotenv
pytest
pytest-asyncio

